Don't allocate for FileStatusRequest
Prior to this commit FileStatusRequest required a
builder API to create. This necessitated a memory
allocation to achieve. This commit alters this API
to solely use a `value class` and Set arithmetic
to arrive at the same end result without a memory
allocation.

This is similar in spirit to Swift's `OptionSet`.