Poll completion queue events in the current thread
Prior to this commit the worker that polled completion queue events was
executed in a separate thread private to a `URing`. This introduced
problems, particuarly around cancellation. The 'prompt cancellation
guarantee' could not be honored because the underlying thread-based
polling mechanism could not be interrupted.

We originally coped with this with a crude 100ms timeout that forced
a cancellation check before waiting for new events again. This was slow.
The 16 unit tests currently in the test suite took ~700ms to run on my
laptop.

This commit removes the need for an additional thread to poll, and
instead uses just another coroutine to do the polling work. We
synchronously check to see if a completion queue event is available
without blocking. If not we `yield` the current coroutine so that other
tasks may execute and continue the synchronous checks at a later time.

The same 16 tests mentioned above now execue in ~7ms instead of ~700ms.
a 100x speedup.